+++ src/fife/gui/fifechan/base/opengl/opengl_gui_graphics.cpp
@@ -31,12 +31,12 @@ namespace FIFE
         m_renderbackend(dynamic_cast<RenderBackendOpenGL*>(RenderBackend::instance()))
     {
         mColor = fcn::Color(255, 255, 255, 255);
-        setTargetPlane(m_renderbackend->getWidth(), m_renderbackend->getHeight());
+        setTargetPlane(static_cast<int>(m_renderbackend->getWidth()), static_cast<int>(m_renderbackend->getHeight()));
     }
 
     void OpenGLGuiGraphics::updateTarget()
     {
-        setTargetPlane(m_renderbackend->getWidth(), m_renderbackend->getHeight());
+        setTargetPlane(static_cast<int>(m_renderbackend->getWidth()), static_cast<int>(m_renderbackend->getHeight()));
     }
 
     void OpenGLGuiGraphics::drawImage(
@@ -115,7 +115,7 @@ namespace FIFE
         m_renderbackend->drawThickLine(
             Point(x1 + top.xOffset, y1 + top.yOffset),
             Point(x2 + top.xOffset, y2 + top.yOffset),
-            width,
+            static_cast<uint8_t>(width),
             mColor.r,
             mColor.g,
             mColor.b,
@@ -130,7 +130,7 @@ namespace FIFE
         for (; it != points.end(); ++it) {
             npoints.emplace_back((*it).x + top.xOffset, (*it).y + top.yOffset);
         }
-        m_renderbackend->drawPolyLine(npoints, width, mColor.r, mColor.g, mColor.b, mColor.a);
+        m_renderbackend->drawPolyLine(npoints, static_cast<uint8_t>(width), mColor.r, mColor.g, mColor.b, mColor.a);
     }
 
     void OpenGLGuiGraphics::drawBezier(fcn::PointVector const & points, int32_t steps, uint32_t width)
@@ -141,7 +141,8 @@ namespace FIFE
         for (; it != points.end(); ++it) {
             npoints.emplace_back((*it).x + top.xOffset, (*it).y + top.yOffset);
         }
-        m_renderbackend->drawBezier(npoints, steps, width, mColor.r, mColor.g, mColor.b, mColor.a);
+        m_renderbackend->drawBezier(
+            npoints, steps, static_cast<uint8_t>(width), mColor.r, mColor.g, mColor.b, mColor.a);
     }
 
     void OpenGLGuiGraphics::drawRectangle(fcn::Rectangle const & rectangle)
@@ -149,8 +150,8 @@ namespace FIFE
         fcn::ClipRectangle const & top = mClipStack.top();
         m_renderbackend->drawRectangle(
             Point(rectangle.x + top.xOffset, rectangle.y + top.yOffset),
-            rectangle.width,
-            rectangle.height,
+            static_cast<uint16_t>(rectangle.width),
+            static_cast<uint16_t>(rectangle.height),
             mColor.r,
             mColor.g,
             mColor.b,
@@ -162,8 +163,8 @@ namespace FIFE
         fcn::ClipRectangle const & top = mClipStack.top();
         m_renderbackend->fillRectangle(
             Point(rectangle.x + top.xOffset, rectangle.y + top.yOffset),
-            rectangle.width,
-            rectangle.height,
+            static_cast<uint16_t>(rectangle.width),
+            static_cast<uint16_t>(rectangle.height),
             mColor.r,
             mColor.g,
             mColor.b,
